fix popup style
